refactor: Move `upstream` package to internal and update imports
- Relocated `upstream` package from `helmtools/` to `internal/`.
- Updated import paths across all affected packages and modules.
- Removed redundant `upstream` references from `helmtools/doc.go`.
